# docs/openapi-v1.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "NotImplementedResponse": {
        "additionalProperties": false,
        "description": "Body returned by a route whose backing service is not built yet.\n\nThese routes are registered rather than omitted so that the OpenAPI document\nis the complete v1 surface: the frontend can be written against the whole API\nfrom day one and build only the parts that answer. The alternative \u2014\npublishing only the built half \u2014 means the contract changes under the\nfrontend halfway through the project, which is the failure mode a contract\nfreeze exists to prevent.\n\n`service` names what is missing and `phase` names when it lands, so nobody has\nto guess whether a 503 is a bug or a scheduled gap.",
        "properties": {
          "endpoint": {
            "title": "Endpoint",
            "type": "string"
          },
          "phase": {
            "description": "Plan phase that delivers it",
            "title": "Phase",
            "type": "string"
          },
          "reason": {
            "default": "This endpoint is part of the frozen v1 contract but its backing ",
            "title": "Reason",
            "type": "string"
          },
          "service": {
            "description": "Service module not yet written",
            "title": "Service",
            "type": "string"
          },
          "status": {
            "default": 503,
            "title": "Status",
            "type": "integer"
          }
        },
        "required": [
          "endpoint",
          "service",
          "phase"
        ],
        "title": "NotImplementedResponse",
        "type": "object"
      },
...
      "ReviewPackGenerateIn": {
        "additionalProperties": false,
        "description": "`POST /review-packs`.",
        "properties": {
          "control_severity_strata": {
            "items": {
              "$ref": "#/components/schemas/Severity"
            },
            "title": "Control Severity Strata",
            "type": "array"
          },
          "entity_ids": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "description": "None means all entities in scope",
            "title": "Entity Ids"
          },
          "max_per_analyst": {
            "default": 3,
            "maximum": 50.0,
            "minimum": 1.0,
            "title": "Max Per Analyst",
            "type": "integer"
          },
          "max_per_cluster": {
            "default": 3,
            "maximum": 50.0,
            "minimum": 1.0,
            "title": "Max Per Cluster",
            "type": "integer"
          },
          "n_control": {
            "default": 10,
            "maximum": 500.0,
            "minimum": 0.0,
            "title": "N Control",
            "type": "integer"
          },
          "n_target": {
            "default": 20,
            "maximum": 500.0,
            "minimum": 1.0,
            "title": "N Target",
            "type": "integer"
          },
          "period_end": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Period End"
          },
          "period_start": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Period Start"
          },
          "seed": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "description": "Recorded on the pack so the sample is reproducible",
            "title": "Seed"
          }
        },
        "title": "ReviewPackGenerateIn",
        "type": "object"
      },
...
      "Severity": {
        "enum": [
          "CRITICAL",
          "HIGH",
          "MEDIUM",
          "LOW"
        ],
        "title": "Severity",
        "type": "string"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
      "Verdict": {
        "description": "Examiner verdict for calibration.",
        "enum": [
          "confirmed",
          "benign",
          "insufficient_information"
        ],
        "title": "Verdict",
        "type": "string"
      },
      "VerdictIn": {
        "additionalProperties": false,
        "description": "`POST /verdicts`. Ledgered.",
        "properties": {
          "case_id": {
            "title": "Case Id",
            "type": "string"
          },
          "evidence_seen": {
            "description": "Row IDs the examiner actually checked",
            "items": {
              "type": "string"
            },
            "title": "Evidence Seen",
            "type": "array"
          },
          "notes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes"
          },
          "pack_id": {
            "title": "Pack Id",
            "type": "string"
          },
          "verdict": {
            "$ref": "#/components/schemas/Verdict"
          }
        },
        "required": [
          "pack_id",
          "case_id",
          "verdict"
        ],
        "title": "VerdictIn",
        "type": "object"
      }
    }
  },
  "info": {
    "description": "Supervisory Analytics Tool for SOC Assessment (SAT-SA) for NCIIPC.\n\nOutputs are **prioritised hypotheses with evidence**, never verdicts or\ncompliance grades. `Not assessable` is a first-class outcome, distinct from\n\"low risk\".\n\nFully offline: no cloud APIs, no CDN, no generative AI in the decision path.\n\n## Reading this contract\n\n`Not assessable` is a **value**, not an absence. Every score field is nullable\nand `null` means \"could not be assessed\" \u2014 never \"scored zero\". The frontend must\nnot coalesce `null` to `0`; a dimension with no data and a dimension with a\nmeasured score of zero are different claims.\n\n`POST /verdicts` records **a supervisor's own conclusion**, written by a human.\nNo Orion-derived payload in this API contains a verdict, a compliance grade, or a\npass/fail.\n\nRoutes marked `503` are part of this frozen contract but their backing service is\nnot written yet. The response body names the service and the phase that delivers\nit. They are listed rather than omitted so the surface does not change shape\nmid-project; `GET /api/v1` splits them into `live` and `pending`.",
    "title": "Orion (SAT-SA) Core Backend",
    "version": "2.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/review-packs": {
      "get": {
        "description": "In the frozen contract; 503 until Phase 11. A pack is a *sample* chosen for human review \u2014 it must keep P(evidence | in pack) equal to P(evidence | population), which is why selection is stratified and randomised rather than ranked by score. A pack built by taking the highest-scoring entities would make every examiner conclusion about it unusable as evidence about the population.",
        "operationId": "list_packs_api_v1_review_packs_get",
        "parameters": [
          {
            "in": "header",
            "name": "X-Actor",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Actor"
            }
          },
          {
            "in": "header",
            "name": "X-Role",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Role"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NotImplementedResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NotImplementedResponse"
                }
              }
            },
            "description": "Not built yet"
          }
        },
        "summary": "Review packs, newest first",
        "tags": [
          "review packs"
        ],
        "x-orion": {
          "phase": "Phase 11 (2.11)",
          "service": "review_pack",
          "status": "pending"
        }
      },
      "post": {
        "description": "In the frozen contract; 503 until Phase 11. Ledgered because the pack is the sampling decision, and a sample that cannot be shown to have been drawn without reference to the scores is the one artefact that would make every conclusion drawn from it contestable.",
        "operationId": "create_pack_api_v1_review_packs_post",
        "parameters": [
          {
            "in": "header",
            "name": "X-Actor",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Actor"
            }
          },
          {
            "in": "header",
            "name": "X-Role",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Role"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReviewPackGenerateIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NotImplementedResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NotImplementedResponse"
                }
              }
            },
            "description": "Not built yet"
          }
        },
        "summary": "Create a review pack. Ledgered.",
        "tags": [
          "review packs"
        ],
        "x-orion": {
          "phase": "Phase 11 (2.11)",
          "service": "review_pack",
          "status": "pending"
        }
      }
    },
    "/api/v1/review-packs/{pack_id}": {
      "get": {
        "description": "In the frozen contract; 503 until Phase 11. Carries the stratum, the selection probabilities (\u03c0), the controls and the hypotheses, so that `fair_selection` \u2014 P(evidence | in pack) = P(evidence | population) \u2014 is checkable from the response rather than asserted by the tool that built the pack.",
        "operationId": "get_pack_api_v1_review_packs__pack_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "pack_id",
            "required": true,
            "schema": {
              "title": "Pack Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-Actor",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Actor"
            }
          },
          {
            "in": "header",
            "name": "X-Role",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Role"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NotImplementedResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NotImplementedResponse"
                }
              }
            },
            "description": "Not built yet"
          }
        },
        "summary": "One pack with its stratum, controls and hypotheses",
        "tags": [
          "review packs"
        ],
        "x-orion": {
          "phase": "Phase 11 (2.11)",
          "service": "review_pack",
          "status": "pending"
        }
      }
    },
    "/api/v1/review-packs/{pack_id}/export": {
      "get": {
        "description": "503 until Phase 13 (2.13). The export is the artefact a supervisor leaves the building with, so it carries the same evidence rows and caveats as the screen rather than a summary of them.",
        "operationId": "export_pack_api_v1_review_packs__pack_id__export_get",
        "parameters": [
          {
            "in": "path",
            "name": "pack_id",
            "required": true,
            "schema": {
              "title": "Pack Id",
              "type": "string"
            }
          },
          {
            "description": "pdf, docx, md or json",
            "in": "query",
            "name": "fmt",
            "required": false,
            "schema": {
              "default": "pdf",
              "description": "pdf, docx, md or json",
              "title": "Fmt",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-Actor",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Actor"
            }
          },
          {
            "in": "header",
            "name": "X-Role",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Role"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NotImplementedResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NotImplementedResponse"
                }
              }
            },
            "description": "Not built yet"
          }
        },
        "summary": "Export a pack as a document",
        "tags": [
          "review packs"
        ],
        "x-orion": {
          "phase": "Phase 13 (2.13)",
          "service": "report_generator",
          "status": "pending"
        }
      }
    },
...
    "/api/v1/verdicts": {
      "get": {
        "description": "In the frozen contract; 503 until Phase 11. Read-only by design \u2014 a recorded conclusion is never edited or withdrawn, only superseded by a later one. Each row is attributed to the supervisor who wrote it, and that attribution is part of the record rather than an audit detail.",
        "operationId": "list_verdicts_api_v1_verdicts_get",
        "parameters": [
          {
            "in": "header",
            "name": "X-Actor",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Actor"
            }
          },
          {
            "in": "header",
            "name": "X-Role",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Role"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NotImplementedResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NotImplementedResponse"
                }
              }
            },
            "description": "Not built yet"
          }
        },
        "summary": "Recorded supervisor conclusions",
        "tags": [
          "review packs"
        ],
        "x-orion": {
          "phase": "Phase 11 (2.11)",
          "service": "review_pack",
          "status": "pending"
        }
      },
      "post": {
        "description": "The supervisor's judgement, entered by a person and attributed to them. Orion proposes hypotheses with evidence; it does not decide, and this route is where the human decision is recorded. 503 until Phase 11.",
        "operationId": "record_verdict_api_v1_verdicts_post",
        "parameters": [
          {
            "in": "header",
            "name": "X-Actor",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Actor"
            }
          },
          {
            "in": "header",
            "name": "X-Role",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Role"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VerdictIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NotImplementedResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NotImplementedResponse"
                }
              }
            },
            "description": "Not built yet"
          }
        },
        "summary": "Record a supervisor's own conclusion about a finding",
        "tags": [
          "review packs"
        ],
        "x-orion": {
          "phase": "Phase 11 (2.11)",
          "service": "review_pack",
          "status": "pending"
        }
      }
    },
...
  },
...
}